[#23] refactor : euphony 패키지 생성
