+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Accounttouch"
         tools:targetApi="31">
-        <activity android:name=".EuphonyTestActivity" />
+        <activity android:name=".euphony.EuphonyTestActivity" />
         <activity
             android:name=".MainActivity"
             android:exported="true"
