+++ app/src/main/java/com/euphony/project/account_touch/euphony/EuphonyManager.kt
@@ -1,4 +1,4 @@
-package com.euphony.project.account_touch
+package com.euphony.project.account_touch.euphony
 
 import euphony.lib.receiver.EuRxManager
 import euphony.lib.transmitter.EuTxManager
